Update deps, add gulp, refactor layouts
Updating our submodules:
`hugo-theme-distinct` v1.0.0 -> v2.0.0
`logo-conn` v1.0.0 -> v1.1.0

The update to the theme is a complete revision that causes us to
refactor our CSS and layouts a bit. Also, with the addition of gulp, we
can follow a `src` -> `static` pattern.

`brand.css` and `css.html` removed because `brand.css` has been migrated
to `hugo-theme-distinct`. With that gone, there's no extra CSS to source
on all generated pages.

Favicon has been moved to `logo-conn` so we'll just symlink to it from
now on. It's nice to know that hugo handles symlinks well.

I want canonical URLs for internal links, so `canonifyURLs` is a natural
choice.  `.Site.BaseURL` to `absURL` where we can, because I prefer it.
`canonifyURLs` actually should not be needed when `absURL` is used. The
documentation is unclear on this, but it seems that `relURL` might
output absolute links when `canonifyURLs` is set to true. This merits
more investigation if I ever want to specify relative internal links in
some places.

We're removing the `categories` taxonomy as I only want to go by tags
(at least for now). I feel uneasy about organizing my posts in a
hierarchical manner that comes with categories.